RLSimion: -renamed all the bullet state/action variables to make them coherent -fixed all the scenes for the bullet worlds -quite a lot of changes to list them all
...but it still doesn't work properly, the robot doesn't move in robot-control

+++ tools/OpenGLRenderer/bindings.h
@@ -36,7 +36,7 @@ public:
 		for (auto it = boundObjects.begin(); it != boundObjects.end(); ++it)
 			delete (*it);
 	}
-	unsigned int getNumBoundObjects() const { return boundObjects.size(); }
+	size_t getNumBoundObjects() const { return boundObjects.size(); }
 	BoundObject* getBoundObject(unsigned int i)
 	{
 		if (i<boundObjects.size())
