+++ tools/OpenGLRenderer/xml-load-utils.cpp
@@ -130,7 +130,7 @@ void loadColladaMatrix(const char* pText, Matrix44& outMatrix)
 	//Collada uses column-major matrices:
 	//https://forums.khronos.org/showthread.php/11010-matrix-and-collada
 	const char* p = pText;
-	int charCount = strlen(p);
+	size_t charCount = strlen(p);
 	int col= 0, row= 0;
 	for (int row = 0; row < 4; ++row)
 	{
@@ -138,7 +138,7 @@ void loadColladaMatrix(const char* pText, Matrix44& outMatrix)
 		{
 			outMatrix.set(col, row, atof(p));
 
-			while (p - pText < charCount && *p != ' ') ++p;
+			while ((size_t)(p - pText) < charCount && *p != ' ') ++p;
 			while (*p == ' ') ++p;
 		}
 	}
@@ -148,16 +148,16 @@ void loadColladaMatrix(const char* pText, Matrix44& outMatrix)
 void loadVector3D(const char* pText, Vector3D& outVec)
 {
 	const char* p = pText;
-	int charCount = strlen(p);
+	size_t charCount = strlen(p);
 
 	outVec.setX(atof(p));
-	while (p - pText < charCount && *p != ' ') ++p;
+	while ((size_t)(p - pText) < charCount && *p != ' ') ++p;
 	while (*p == ' ') ++p;
 	outVec.setY(atof(p));
-	while (p - pText < charCount && *p != ' ') ++p;
+	while ((size_t)(p - pText) < charCount && *p != ' ') ++p;
 	while (*p == ' ') ++p;
 	outVec.setZ(atof(p));
-	while (p - pText < charCount && *p != ' ') ++p;
+	while ((size_t)(p - pText) < charCount && *p != ' ') ++p;
 }
 
 unsigned int countChildren(tinyxml2::XMLElement* pParentNode, const char* xmlTag)
